[refactor] renaming sensor to sensor_view
